+++ js/ui.validator.js
@@ -27,13 +27,13 @@ Validator.DEFAULTS = {
   validateOnSubmit: true,
   // Elements to validate with allValid (only validating visible elements)
   // :input: selects all input, textarea, select and button elements.
-  allFields: ':input:visible:not(:button, :disabled, .am-novalidate)',
+  allFields: ':input:visible:not(:submit, :button, :disabled, .am-novalidate)',
 
   // Custom events
   customEvents: 'validate',
 
   // Keyboard events
-  keyboardFields: ':input:not(:button, :disabled,.am-novalidate)',
+  keyboardFields: ':input:not(:submit, :button, :disabled, .am-novalidate)',
   keyboardEvents: 'focusout, change', // keyup, focusin
 
   activeKeyup: true,
@@ -81,6 +81,8 @@ Validator.DEFAULTS = {
   submit: null
 };
 
+Validator.VERSION = '2.0.0';
+
 /* jshint -W101 */
 Validator.patterns = {
   email: /^((([a-zA-Z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-zA-Z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-zA-Z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-zA-Z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-zA-Z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-zA-Z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-zA-Z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-zA-Z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-zA-Z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-zA-Z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.?$/,
@@ -96,7 +98,7 @@ Validator.patterns = {
 
 Validator.validationMessages = {
   zh_CN: {
-    valueMissing: '请填写此字段',
+    valueMissing: '请填写（选择）此字段',
     customError: {
       tooShort: '至少填写 %s 个字符',
       checkedOverflow: '至多选择 %s 项',
@@ -109,6 +111,15 @@ Validator.validationMessages = {
     tooLong: '至多填写 %s 个字符',
     typeMismatch: '请按照要求的类型填写'
   }
+};
+
+Validator.ERROR_MAP = {
+  tooShort: 'minlength',
+  checkedOverflow: 'maxchecked',
+  checkedUnderflow: 'minchecked',
+  rangeOverflow: 'max',
+  rangeUnderflow: 'min',
+  tooLong: 'maxlength'
 };
 
 // TODO: 考虑表单元素不是 form 子元素的情形
@@ -148,12 +159,32 @@ Validator.prototype.init = function() {
   });
 
   $element.submit(function(e) {
+    // user custom submit handler
     if (typeof options.submit === 'function') {
       return options.submit.call(_this, e);
     }
 
     if (options.validateOnSubmit) {
-      return _this.isFormValid();
+      var formValidity = _this.isFormValid();
+
+      // sync validate, return result
+      if ($.type(formValidity) === 'boolean') {
+        return formValidity;
+      }
+
+      if ($element.data('amui.checked')) {
+        return true;
+      } else {
+        $.when(formValidity).then(function() {
+          // done, submit form
+          $element.data('amui.checked', true).submit();
+        }, function() {
+          // fail
+          $element.data('amui.checked', false).
+            find('.' + options.inValidClass).eq(0).focus();
+        });
+        return false;
+      }
     }
   });
 
@@ -179,6 +210,15 @@ Validator.prototype.init = function() {
   // active filed
   bindEvents('.am-active', 'keyup', 50);
   bindEvents('textarea[maxlength]', 'keyup', 50);
+
+  /*if (options.errorMessage === 'tooltip') {
+    this.$tooltip = $('<div></div>', {
+      'class': 'am-validator-message',
+      id: UI.utils.generateGUID('am-validator-message')
+    });
+
+    $(document.body).append(this.$tooltip);
+  }*/
 };
 
 Validator.prototype.isValid = function(field) {
@@ -207,6 +247,7 @@ Validator.prototype.validate = function(field) {
   var $radioGroup = null;
   var $checkboxGroup = null;
   // if checkbox, return `:chcked` length
+  // NOTE: checkbox and radio should have name attribute
   var value = ($field.is('[type=checkbox]')) ?
     ($checkboxGroup = $element.find('input[name="' + field.name + '"]')).
       filter(':checked').length : ($field.is('[type=radio]') ?
@@ -216,6 +257,7 @@ Validator.prototype.validate = function(field) {
   // if checkbox, valid the first input of checkbox group
   $field = ($checkboxGroup && $checkboxGroup.length) ?
     $checkboxGroup.first() : $field;
+
   var required = ($field.attr('required') !== undefined) &&
     ($field.attr('required') !== 'false');
   var maxLength = parseInt($field.attr('maxlength'), 10);
@@ -305,12 +347,17 @@ Validator.prototype.validate = function(field) {
   };
 
   // Run custom validate
+  // NOTE: async custom validate should return Deferred project
   var customValidate;
   (typeof options.validate === 'function') &&
     (customValidate = options.validate.call(this, validity));
 
+  // Deferred
   if (customValidate) {
+    var dfd = new $.Deferred();
+    $field.data('amui.dfdValidity', dfd.promise());
     return $.when(customValidate).always(function(validity) {
+      dfd[validity.valid ? 'resolve' : 'reject'](validity);
       validateComplete.call(_this, validity);
     });
   }
@@ -325,7 +372,7 @@ Validator.prototype.markField = function(validity) {
 };
 
 // check all fields in the form are valid
-Validator.prototype.validateAll = function() {
+Validator.prototype.validateForm = function() {
   var _this = this;
   var $element = this.$element;
   var options = this.options;
@@ -334,11 +381,14 @@ Validator.prototype.validateAll = function() {
   var valid = true;
   var formValidity = [];
   var $inValidFields = $([]);
+  var promises = [];
+  // for async validate
+  var async = false;
 
   $element.trigger('validate.form.validator.amui');
 
   // Filter radio with the same name and keep only one,
-  // since they will be checked as a group by isValid()
+  //   since they will be checked as a group by validate()
   var $filteredFields = $allFields.filter(function(index) {
     var name;
     if (this.tagName === 'INPUT' && this.type === 'radio') {
@@ -352,18 +402,38 @@ Validator.prototype.validateAll = function() {
   });
 
   $filteredFields.each(function() {
+    var $this = $(this);
     var fieldValid = _this.isValid(this);
+    var fieldValidity = $this.data('validity');
+
     valid = !!fieldValid && valid;
-    formValidity.push($(this).data('validity'));
+    formValidity.push(fieldValidity);
+
     if (!fieldValid) {
       $inValidFields = $inValidFields.add($(this), $element);
     }
+
+    // async validity
+    var promise = $this.data('amui.dfdValidity');
+
+    if (promise) {
+      promises.push(promise);
+      async = true;
+    } else {
+      // convert sync validity to Promise
+      var dfd = new $.Deferred();
+      promises.push(dfd.promise());
+      dfd[fieldValid ? 'resolve' : 'reject'](fieldValidity);
+    }
   });
 
+  // NOTE: If there are async validity, the valid may be not exact result.
   var validity = {
     valid: valid,
     $invalidFields: $inValidFields,
-    validity: formValidity
+    validity: formValidity,
+    promises: promises,
+    async: async
   };
 
   $element.trigger('validated.form.validator.amui', validity);
@@ -372,14 +442,34 @@ Validator.prototype.validateAll = function() {
 };
 
 Validator.prototype.isFormValid = function() {
-  var formValid = this.validateAll();
-  if (!formValid.valid) {
-    formValid.$invalidFields.first().focus();
-    this.$element.trigger('invalid.validator.amui');
-    return false;
+  var _this = this;
+  var formValidity = this.validateForm();
+  var triggerValid = function(type) {
+    _this.$element.trigger(type + '.validator.amui');
+  };
+
+  if (formValidity.async) {
+    var masterDfd = new $.Deferred();
+
+    $.when.apply(null, formValidity.promises).then(function() {
+      masterDfd.resolve();
+      triggerValid('valid');
+    }, function() {
+      masterDfd.reject();
+      triggerValid('invalid');
+    });
+
+    return masterDfd.promise();
+  } else {
+    if (!formValidity.valid) {
+      formValidity.$invalidFields.first().focus();
+      triggerValid('invalid');
+      return false;
+    }
+
+    triggerValid('valid');
+    return true;
   }
-  this.$element.trigger('valid.validator.amui');
-  return true;
 };
 
 // customErrors:
@@ -400,6 +490,48 @@ Validator.prototype.createValidity = function(validity) {
     // Returns true if the element has no value but is a required field
     valueMissing: validity.valueMissing || false
   }, validity);
+};
+
+Validator.prototype.getValidationMessage = function(validity) {
+  var messages = Validator.validationMessages[this.options.locales];
+  var error;
+  var message;
+  var placeholder = '%s';
+  var $field = $(validity.field);
+
+  if ($field.is('[type="checkbox"]') || $field.is('[type="radio"]')) {
+    $field = this.$element.find('[name=' + $field.attr('name') + ']').first();
+  }
+
+  // get error name
+  $.each(validity, function(key, val) {
+    // skip `field` and `valid`
+    if (key === 'field' || key === 'valid') {
+      return key;
+    }
+
+    // Amaze UI custom error type
+    if (key === 'customError' && val) {
+      error = val;
+      messages = messages.customError;
+      return false;
+    }
+
+    // W3C specs error type
+    if (val === true) {
+      error = key;
+      return false;
+    }
+  });
+
+  message = messages[error] || undefined;
+
+  if (message && Validator.ERROR_MAP[error]) {
+    message = message.replace(placeholder,
+      $field.attr(Validator.ERROR_MAP[error]) || '规定的');
+  }
+
+  return message;
 };
 
 function Plugin(option) {