+++ widget/menu/src/menu.js
@@ -148,6 +148,6 @@ $(function() {
 });
 
 module.exports = $.AMUI.menu = {
-  VERSION: '4.0.2',
+  VERSION: '4.0.3',
   init: menuInit
 };